--- back/dubeng-user/src/main/java/com/ssafy/dubenguser/dto/Res.java
@@ -1,11 +1,0 @@
-package com.ssafy.dubenguser.dto;
-
-import lombok.Data;
-
-@Data
-public class Res<T> {
-    private int statusCode;
-    private T Data;
-
-    private String message;
-}
