--- back/dubeng-user/src/main/java/com/ssafy/dubenguser/dto/Token.java
@@ -1,11 +1,0 @@
-package com.ssafy.dubenguser.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Token {
-    private String accessToken;
-    private String refreshToken;
-}
