--- back/dubeng-user/src/main/java/com/ssafy/dubenguser/dto/UserCalenderReq.java
@@ -1,9 +1,0 @@
-package com.ssafy.dubenguser.dto;
-
-import com.ssafy.dubenguser.entity.UserCalendar;
-import lombok.Data;
-
-@Data
-public class UserCalenderReq {
-
-}
