+++ back/dubeng-user/src/main/java/com/ssafy/dubenguser/service/UserService.java
@@ -5,9 +5,7 @@ import com.ssafy.dubenguser.dto.*;
 import java.util.List;
 
 public interface UserService {
-    public void addUser(UserJoinReq requestDTO, String accessToken);
-    public boolean checkEnrolledMember(String id);
-    public boolean checkExistNickname(String nickname);
+
     public UserProfileRes findProfile(String accessToken);
     public UserCalendarRes findCalendar(String accessToken);
     public List<UserRecordRes> findRecord(UserRecordReq request);
